Remove repeat size
This option doesn't fit well with the redesigned UI; the user will
surely not go to the app menu to choose from two competing options for
starting a new game, when there is a huge Start Over button in the
window itself. It's too inconveinent for a conveience option.

(It's also slightly confusing.)