Updated get-help.page according to UI change for 3.12
The penalty has changed from ten seconds each time to ten seconds on the first hint
and then additional ten seconds progressively on subsequent hints till the
ninth hint. The string "Warn if too many flags are placed" has been changed to
"Warm if too many flags are placed next to a number" in the preference dialog
and so in the main window.

https://bugzilla.gnome.org/show_bug.cgi?id=725985