Rework minefield sizing and aspect ratio calculations
There is a circular dependency between the minefiled_aspect ratio
calculations and the minefiled + aspect_child + button_box allocations
which results in the initial minefield often drawn with bad proportions.

Attempt to workaround that by scheduling one addition resize after the
first cycle of allocations on a new minefiled configuration.