Sidebar buttons moved to headerbar
In the existing layout sidebar buttons (new game, pause, clear board and
high scores) were taking big amount of screen space. When window had
a vertical aspect ratio more space was wasted above and below the
minefield. The loss of space was especially painfull on a smartphone
screen.

This commit introduces changes in layout to minimize the loss of screen
space:
- new game, pause and clear board buttons were moved to the headerbar
- high scores button was removed (the action remains available in the
  primary menu)
- clock and mines counter may be moved below the minefield if the window
  has vertical aspect ratio

Semi-related changes:
- primary menu definition was moved from vala to interface.ui
- minimum tile size was increased to 30