Add D-Bus service file
And mark the application as D-Bus activatable. This allows application
launchers to activate it via D-Bus.

Update the action names to align with the ActionEntry names, and make them
work even after the application is launched: when the user choose a board
size during a game is in progress, it will be asked if replace the current
game.