Fix invalid cast from 'MineWindow' to 'AdwWindow'
Running Mines in jhbuild, I was surprised to notice it crash on startup.
I have G_DEBUG=fatal-criticals set in my jhbuild environment, and Mines
is performing an invalid cast when creating the UI. Unfortunately
AdwApplicationWindow is confusingly not a subclass of AdwWindow.
Luckily, although it's stored in the wrong type all over the place, it's
apparently never *used* improperly, so the critical is the worst that
happens.

Fixes "invalid cast from 'MineWindow' to 'AdwWindow'"